refactor: Rename search screen package
