+++ presentation/src/main/java/com/london/presentation/screen/search/SearchCategory.kt
@@ -1,4 +1,4 @@
-package com.london.presentation.screens.search_screen
+package com.london.presentation.screen.search
 
 enum class SearchCategory(val title: String) {
     Movies("Movies"),
